requirements.txt
aioredis==1.3.1
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
arrow==0.17.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dishka==1.5.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
hiredis==3.1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jwt==1.3.1
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pyasn1==0.4.8
pycasestyle==0.1.4
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
redis==6.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
Unidecode==1.3.8
uvicorn==0.34.2